requirements.txt
openai
ffmpeg-python
python-dotenv 